Noticed that load shed where there is no load

+++ scripts/main_model_runs.jl
@@ -3,12 +3,17 @@ sn = script_name(@__FILE__)
 mkrootdirs(plotsdir(sn))
 
 function res_act_net(opts)
-    return opts["reserve_shedding_limit"] <= 1.0 && (
+    return (
+        opts["copperplate"] == false &&
         isempty(opts["upward_reserve_levels_included_in_redispatch"]) ==
         false ||
         isempty(opts["downward_reserve_levels_included_in_redispatch"]) ==
         false
     )
+end
+
+function has_OR(gep::GEPM)
+    return isnothing(gep_1[:I, :uncertainty]) == false
 end
 
 opts = options_diff_days(sn)[4]
@@ -75,7 +80,7 @@ opts_vec = vcat(
                 "downward_reserve_levels_included_in_redispatch" => 1:10,
                 "save_path" => "$(opts["save_path"])_L⁺=1:10_L⁻=1:10",
             ),
-        ) for opts in opts_vec if opts["reserve_shedding_limit"] < 1.0
+        ) for opts in opts_vec if opts["reserve_shedding_limit"] <= 1.0
     ]...,
 )
 opts_vec = vcat(
@@ -103,17 +108,20 @@ opts_vec = vcat(
     ],
 )
 
-function main_model_run(opts; make_plots=true)
+function main_model_run(opts; make_plots=false)
     gep = run_GEPPR(opts)
 
-    if make_plots
+    dis_file = joinpath(opts["save_path"], "dispatch.png")
+    res_file = joinpath(opts["save_path"], "ls_and_rs.png")
+
+    if make_plots && (isfile(dis_file) == false && isfile(res_file) == false)
         # Plot dispatch
         plt = plot_dispatch(gep, 1)
-        Plots.savefig(plt, joinpath(opts["save_path"], "dispatch.png"))
+        Plots.savefig(plt, dis_file)
 
         # If has reserves, plot these
         if opts["operating_reserves_type"] != "none"
-            apply_operating_reserves!(gep, opts)
+            has_OR(gep) == false && apply_operating_reserves!(gep, opts)
             plt = plot_reserves_simple(gep, 1)
             Plots.savefig(plt, joinpath(opts["save_path"], "reserves.png"))
         end
@@ -134,16 +142,18 @@ function main_model_run(opts; make_plots=true)
             ylabel="Power",
             ylims=(0, max(100, maximum(ls))),
         )
-        Plots.savefig(plt, joinpath(opts["save_path"], "ls_and_rs.png"))
+        Plots.savefig(plt, res_file)
     end
 
     return gep
 end
 
+main_model_run(opts_vec[63])
+
 gep_vec = GEPM[]
 for opts in opts_vec
     try
-        push!(gep_vec, main_model_run(opts; make_plots=false))
+        push!(gep_vec, main_model_run(opts; make_plots=true))
     catch e
         @warn "$(opts["save_path"]) failed."
         println(string(e))
@@ -183,25 +193,152 @@ function analyse_main_model_runs(gep_vec, opts_vec)
         df_no_OR = filter(row -> row["OR"] == false, df)
         CSV.write(joinpath(path, "no_OR.csv"), df_no_OR)
 
-        df_no_RANet = filter(row -> row["OR"] == true && row["RANet"] == false && row["AbsImb"] == false, df)
-        DataFrames.sort!(df_no_RANet, ["UC", "DANet", "PSCD"])
+        df_no_RANet = filter(
+            row ->
+                row["OR"] == true &&
+                    row["RANet"] == false &&
+                    row["AbsImb"] == false,
+            df,
+        )
+        DataFrames.sort!(df_no_RANet, ["UC", "DANet", "PSCD", "RSV"])
         CSV.write(joinpath(path, "no_RANet.csv"), df_no_RANet)
 
-        df_no_AbsImb = filter(row -> row["OR"] == true && row["RANet"] == true && row["AbsImb"] == false, df)
-        DataFrames.sort!(df_no_AbsImb, ["UC", "DANet", "PSCD"])
+        df_no_AbsImb = filter(
+            row ->
+                row["OR"] == true &&
+                    row["RANet"] == true &&
+                    row["AbsImb"] == false,
+            df,
+        )
+        DataFrames.sort!(df_no_AbsImb, ["UC", "DANet", "PSCD", "RSV"])
         CSV.write(joinpath(path, "no_AbsImb.csv"), df_no_AbsImb)
 
-        df_no_ConvImb = filter(row -> row["OR"] == true && row["RANet"] == true && row["AbsImb"] == true, df)
-        DataFrames.sort!(df_no_AbsImb, ["UC", "DANet", "PSCD"])
+        df_no_ConvImb = filter(
+            row ->
+                row["OR"] == true &&
+                    row["RANet"] == true &&
+                    row["AbsImb"] == true,
+            df,
+        )
+        DataFrames.sort!(df_no_ConvImb, ["UC", "DANet", "PSCD", "RSV"])
         CSV.write(joinpath(path, "no_ConvImb.csv"), df_no_ConvImb)
 
         return nothing
     end
 
-    analyse_effect_of_constraints(filter(row -> ismissing(row["LM"]), df), joinpath(opts["save_path"], "LM=1.0"))
-    analyse_effect_of_constraints(filter(row -> ismissing(row["LM"]) == 1.5, df), joinpath(opts["save_path"], "LM=1.5"))
+    analyse_effect_of_constraints(
+        filter(row -> ismissing(row["LM"]), df),
+        joinpath(opts["save_path"], "LM=1.0"),
+    )
+    analyse_effect_of_constraints(
+        filter(row -> ismissing(row["LM"]) == false && row["LM"] == 1.5, df),
+        joinpath(opts["save_path"], "LM=1.5"),
+    )
 
     return df
 end
 
 df = analyse_main_model_runs(gep_vec, opts_vec)
+
+# Why does reserve shedding happen?
+function how_is_reserve_shed_without_increasing_load_shedding(gep_vec, opts_vec)
+    # When is load and reserve shed?
+    i = 7
+    ls_0_5 = dropdims(
+        sum(gep_vec[i + 4][:loadShedding].data; dims=(2, 3)); dims=(2, 3)
+    )
+    rs_0_5 = dropdims(
+        sum(gep_vec[i + 4][:rsL⁺].data; dims=(2, 3, 4)); dims=(2, 3, 4)
+    )
+    @show ls_0_5_idx = findall(ls_0_5 .> 1e-3)
+    @show rs_0_5_idx = findall(rs_0_5 .> 1e-6)
+
+    # Load shedding occurs same time when RSL = 0 and RSL = 0.5
+    ls_0 = dropdims(
+        sum(gep_vec[i][:loadShedding].data; dims=(2, 3)); dims=(2, 3)
+    )
+    ls_0_idx = findall(ls_0_5 .> 1e-3)
+    @show ls_0_idx == ls_0_5_idx
+
+    # Plot load shedding
+    @show sum(ls_0[:, 10] .- ls_0_5[:, 10])
+    N = GEPPR.get_set_of_nodes(gep_vec[i])
+    x = eachindex(N)
+    display(
+        Plots.bar(
+            hcat(x...),
+            ls_0_5[:, 10]';
+            lab="",
+            ylabel="Load shedding [MWh]",
+            xticks=(x, N),
+            tickfontrotation=90,
+            margin=10mm,
+            size=(1200, 800),
+        ),
+    )
+
+    # When is commitment different in both cases?
+    z_0_5 = dropdims(gep_vec[i + 4][:z].data; dims=(2, 3))
+    z_0 = dropdims(gep_vec[i][:z].data; dims=(2, 3))
+    @show findall(z_0_5 .- z_0 .!= 0)
+
+    # When is generation different in both cases?
+    q_0_5 = dropdims(gep_vec[i + 4][:q].data; dims=(2, 3))
+    q_0 = dropdims(gep_vec[i][:q].data; dims=(2, 3))
+    @show q_idx = findall(abs.(q_0_5[:, 9] .- q_0[:, 9]) .> 1e-3)
+    GN = GEPPR.get_set_of_nodal_generators(gep_vec[i])
+    @show GN[q_idx]
+end
+
+# how_is_reserve_shed_without_increasing_load_shedding(gep_vec, opts_vec)
+
+function load_shedding_distribution(gep_vec, opts_vec, idx)
+    plotly()
+    gep = gep_vec[63]
+
+    ls = dropdims(gep[:loadShedding].data; dims=(2, 3))
+    plt = Plots.plot(
+        sum(ls; dims=1)[:]; ylab="Load shedding [MWh]", xlab="Time [h]", lab=""
+    )
+    Plots.savefig(plt, plotsdir(sn, "ls_timeseries_idx=$(idx).png"))
+
+    N = GEPPR.get_set_of_nodes(gep)
+    x = eachindex(N)
+    plt = Plots.bar(
+        hcat(x...),
+        sum(ls; dims=2)';
+        ylab="Load shedding / Load [-]",
+        xticks=(x, N),
+        lab="",
+        tickfontrotation=90,
+        bw=1.0,
+        lw=0.0,
+        c=:blue,
+    )
+    Plots.savefig(plt, plotsdir(sn, "ls_per_node_idx=$(idx).png"))
+
+    D = dropdims(GEPPR.get_demand(gep); dims=(2, 3))
+
+    norm_ls = (sum(ls; dims=2) ./ sum(D; dims=2))'
+    plt = Plots.bar(
+        hcat(x...),
+        norm_ls;
+        ylab="Load shedding / Load [-]",
+        xticks=(x, N),
+        lab="",
+        tickfontrotation=90,
+        bw=1.0,
+        lw=0.0,
+        c=:blue,
+    )
+    x_inf = [i for i in eachindex(norm_ls[:]) if isinf(norm_ls[i])]
+    Plots.scatter!(
+        plt,
+        x_inf,
+        [0.0 for i in eachindex(x_inf)];
+        markershape=:diamond,
+        markercolor=:red,
+        markersize=4,
+    )
+    return Plots.savefig(plt, plotsdir(sn, "ls_per_node_idx=$(idx).png"))
+end
